+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version History
             devices since `v2.3.0`
         -   Fixed not calling the callback anymore after requesting cancellation,
             while the operation is still being executed
+-   Added support for creating shared buffers on Metal devices
 
 ### Changes in v2.3.0: